final styling of event weather
