+++ src/index.html
@@ -23,6 +23,7 @@
     <article class="underHeader">
         <article class="dashboard--left">
             <div class="weatherContainer"></div>
+            
         </article>
 
         <article class="dashboard--center">
